django/airbnb_clone_coding/pyproject.toml
[tool.poetry]
name = "aribnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "~4.0"
djangorestframework = "^3.14.0"
pyjwt = "^2.6.0"
django-environ = "^0.9.0"
dj-database-url = "^1.2.0"
psycopg2-binary = "^2.9.5"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
